Keep AG-UI browser imports off server logger paths
Studio imports veryfront/chat/ag-ui in a browser bundle. That entrypoint only needs pure streamed tool-input helpers, but those helpers lived in agent/streaming/data-stream.ts beside the server logger. Importing the chat parser therefore pulled server-only modules into Vite browser tests.

This moves the tool-input merge and parse helpers into a pure streaming module, keeps the existing data-stream exports as re-exports, and bumps the package version for the patch release.

Constraint: veryfront/chat/ag-ui must bundle in browser environments without Node async_hooks.

Rejected: Shim node:async_hooks in Studio | keeps the SDK browser entrypoint server-tainted.

Confidence: high

Scope-risk: narrow

Directive: Keep browser-safe chat entrypoints off server logging and runtime modules.

Tested: deno fmt src/agent/streaming/data-stream.ts src/agent/streaming/tool-input.ts src/chat/ag-ui.ts src/chat/ag-ui.test.ts

Tested: deno lint src/agent/streaming/data-stream.ts src/agent/streaming/tool-input.ts src/chat/ag-ui.ts src/chat/ag-ui.test.ts

Tested: deno test --no-check --allow-all --unstable-worker-options --unstable-net src/chat/ag-ui.test.ts src/agent/streaming/data-stream.test.ts

Tested: deno check --lock=deno.lock --frozen src/chat/ag-ui.ts src/agent/streaming/data-stream.ts src/agent/streaming/tool-input.ts src/chat/ag-ui.test.ts src/agent/streaming/data-stream.test.ts

Tested: deno task test:unit --no-lock

Tested: git diff --check

Not-tested: Browser bundle verification against the published npm artifact; pending after publishing 0.1.874.
